IndexingVariantContextWriter cleanup   * extract PositionalOutputStream   * extract setIndexSequenceDictionary to IndexCreator with no-op default and TribbleIndexCreator implementation
